fix(ci): line-limit — extract feature-PR retry helpers under 1500-line gate
Move the cohesive feature/sync PR identity, adoption live-ref sync, trusted
freeze cleanup, and hosted adoption qualification helpers into
workspaces_retry_feature_pr.py so workspaces_retry.py stays well under the
first-party gate with headroom for remaining narrowly justified corrections.
